Discovery: Make room for new implementation
Convert current DiscoveryPage view in a wrapper that conditionally uses
the old or the new one. The new one is just a hello-world page for now.
